* removed obsolete function LaserAlignment::fit and the file plugins/LaserAlignmentFit.cc housing it * LaserAlignment::endRun: fixed wrong assignment of beamId when filling AT beams in TkLasBeamCollection * renamed laserAlignment::getTOBProfileOffset ::getTIBTOBProfileOffset and refined it to max. precision (->changes to LASProfileJudge) * removed obsolete function LaserAlignment::alignmentAlgorithm and file plugins/LaserAlignmentAlignmentAlgorithm.cc housing it * changes in plugins/LaserAlignment.cc due to redefinition of trackbased interface containers * removed .h & .cc files of: AlignmentAlgorithmBW LaserAlignmentAlgorithmNegTEC LaserAlignmentAlgorithmPosTEC   LaserAlignmentAlgorithmTEC2TEC LaserAlignmentNegTEC LaserAlignmentPosTEC LaserAlignmentTEC2TEC (and their traces)

+++ Alignment/LaserAlignment/src/LASPeakFinder.h
@@ -22,7 +22,7 @@ class LASPeakFinder {
 
  public:
   LASPeakFinder();
-  bool FindPeakIn( const LASModuleProfile&, std::pair<double,double>&, const int );
+  bool FindPeakIn( const LASModuleProfile&, std::pair<double,double>&, const double );
 
 };
